requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
tqdm>=4.65.0

# External APIs (optional)
openai>=1.0.0  # For DALL-E API

# ML utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0


